Changed NetSendOrder to send the step as 32 bits rather than 8, so that games can lost longer than 8 seconds

+++ src/NetMessage.cpp
@@ -178,7 +178,7 @@ void NetSendOrder::encodeData(GAGCore::OutputStream* stream) const
 	stream->writeUint8(order->getOrderType(), "orderType");
 	stream->write(order->getData(), order->getDataLength(), "data");
 	stream->writeUint8(order->sender, "sender");
-	stream->writeUint8(order->ustep, "ustep");
+	stream->writeUint32(order->ustep, "ustep");
 	stream->writeLeaveSection();
 }
 
@@ -194,7 +194,7 @@ void NetSendOrder::decodeData(GAGCore::InputStream* stream)
 	
 	order = Order::getOrder(buffer, size);
 	order->sender = stream->readUint8("sender");
-	order->ustep = stream->readUint8("ustep");
+	order->ustep = stream->readUint32("ustep");
 	
 	delete buffer;
 }
